Corrected a bug where the background was not scrolling correctly

+++ Space-Harriers-Clone/ModuleBackground.h
@@ -37,8 +37,8 @@ private:
 	int currentLevelHorizonDecalTextureHeight;
 
 
-	int backgroundTextureOffset;
-	int decalTextureOffset;
+	float backgroundTextureOffset;
+	float decalTextureOffset;
 private:
 
 	void DrawBackground();
